Fix issues in node + move 'query' option from upload to base options
