+++ src/client/web.ts
@@ -1,5 +1,12 @@
 import { downloadFile, downloadFileHns, openFile, openFileHns } from "../download/web";
-import { uploadFile, uploadDirectory, uploadDirectoryRequest, uploadFileRequest } from "../upload/web";
+import {
+  uploadFile,
+  uploadDirectory,
+  uploadDirectoryRequest,
+  uploadFileRequest,
+  uploadFileContent,
+  uploadFileContentRequest,
+} from "../upload/web";
 import { SkynetClient as Client } from "./index";
 
 export class SkynetClient extends Client {
@@ -14,4 +21,6 @@ export class SkynetClient extends Client {
   protected uploadFileRequest = uploadFileRequest;
   uploadDirectory = uploadDirectory;
   protected uploadDirectoryRequest = uploadDirectoryRequest;
+  uploadFileContent = uploadFileContent;
+  protected uploadFileContentRequest = uploadFileContentRequest;
 }
