+++ src/upload/node.ts
@@ -171,6 +171,73 @@ export async function uploadDirectoryFromPathRequest(
 }
 
 /**
+ * Uploads a file to Skynet.
+ *
+ * @param this - SkynetClient
+ * @param fileContents - The file contents to upload.
+ * @param fileName - The desired name for the file.
+ * @param [customOptions] - Additional settings that can optionally be set.
+ * @param [customOptions.endpointPath="/skynet/skyfile"] - The relative URL path of the portal endpoint to contact.
+ * @returns - The returned skyfile information including skylink, merkleroot and bitfield.
+ * @throws - Will throw if the request is successful but the upload response does not contain a complete response.
+ */
+export async function uploadFileContent(
+  this: SkynetClient,
+  fileContents: string,
+  fileName: string,
+  customOptions?: CustomUploadOptions
+): Promise<UploadRequestResponse> {
+  const response = await this.uploadFileContentRequest(fileContents, fileName, customOptions);
+
+  /* istanbul ignore next */
+  if (
+    typeof response.data.skylink !== "string" ||
+    typeof response.data.merkleroot !== "string" ||
+    typeof response.data.bitfield !== "number"
+  ) {
+    throw new Error(
+      "Did not get a complete upload response despite a successful request. Please try again and report this issue to the devs if it persists."
+    );
+  }
+
+  const skylink = formatSkylink(response.data.skylink);
+  const merkleroot = response.data.merkleroot;
+  const bitfield = response.data.bitfield;
+
+  return { skylink, merkleroot, bitfield };
+}
+
+/**
+ * Makes a request to upload a file to Skynet.
+ *
+ * @param this - SkynetClient
+ * @param fileContents - The file contents to upload.
+ * @param fileName - The desired name for the file.
+ * @param [customOptions] - Additional settings that can optionally be set.
+ * @param [customOptions.endpointPath="/skynet/skyfile"] - The relative URL path of the portal endpoint to contact.
+ * @returns - The upload response.
+ */
+export async function uploadFileContentRequest(
+  this: SkynetClient,
+  fileContents: string,
+  fileName: string,
+  customOptions?: CustomUploadOptions
+): Promise<AxiosResponse> {
+  const opts = { ...defaultUploadOptions, ...this.customOptions, ...customOptions };
+
+  const formData = new FormData();
+  formData.append(opts.portalFileFieldname, fileContents, fileName);
+
+  return this.executeRequest({
+    ...opts,
+    method: "post",
+    // @ts-expect-error TS doesn't recognize this external FormData.
+    data: formData,
+    headers: formData.getHeaders(),
+  });
+}
+
+/**
  * Returns the full recursive list of files inside a directory.
  *
  * @param filepath - The directory path.
